Add destroy and delete buttons

+++ app/controllers/posts_controller.rb
@@ -25,17 +25,21 @@ class PostsController < ApplicationController
       render :new
     end
   end
+
   def destroy
-    current_post = Post.find_by(id: params[:post_id])
-    authorize! :destroy, current_post
-    current_post.comments.each(&:destroy)
-    current_post.likes.each(&:destroy)
-    current_post.destroy
-    user_post_counter = User.find_by(id: params[:id])
-    user_post_counter.posts_counter -= 1
-    user_post_counter.save
-    redirect_to posts_path(id: params[:id])
+    current_post = Post.find(params[:id])
+    if current_post
+      authorize! :destroy, current_post
+      current_post.comments.each(&:destroy)
+      current_post.likes.each(&:destroy)
+      current_post.destroy
+      user_post_counter = User.find(params[:user_id])
+      redirect_to user_posts_path(current_user), notice: 'post was successfully deleted.'
+    else
+      redirect_to redirect_url, alert: 'Failed to delete the post.'
+    end
   end
+
   private
 
   def post_params
